Weights are now from 0 to 100 instead of floating point.

+++ disco/types.go
@@ -8,7 +8,7 @@ import (
 type PeerLink struct {
 	Account eos.AccountName `json:"account"`
 	Comment string          `json:"comment"`
-	Weight  float64         `json:"weight"`
+	Weight  uint8           `json:"weight"`
 }
 type ContentRef struct {
 	Name    string `json:"name"`
